requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
absl-py==0.14.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
fsspec==2021.9.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
pkg_resources==0.0.0
pluggy==1.0.0
protobuf==3.18.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyDeprecate==0.3.1
pyparsing==2.4.7
pytest==6.2.5
pytorch-lightning==1.4.8
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
toml==0.10.2
torch==1.9.0+cpu
torchvision==0.10.0+cpu
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
yarl==1.6.3